requirements.txt
alembic==1.7.6
aniso8601==9.0.1
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.18.0b2
fastapi==0.74.1
FastAPI-SQLAlchemy==0.2.1
graphene==3.0
graphql-core==3.1.7
graphql-relay==3.1.5
greenlet==2.0.0a1
h11==0.13.0
idna==3.3
Mako==1.1.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.17.1
starlette-graphene3==0.5.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5